Move extensions to their own file
