+++ extensions/aliases.php
@@ -1,0 +1,4 @@
+<?php
+
+class_alias('Kirby\Kql\Kql', 'Kql');
+class_alias('Kirby\Kql\Interceptor', 'Kirby\Kql\Interceptors\Interceptor');
