+++ extensions/helpers.php
@@ -1,0 +1,8 @@
+<?php
+
+use Kirby\Kql\Kql;
+
+function kql($input, $model = null)
+{
+	return Kql::run($input, $model);
+}
