+++ index.php
@@ -1,37 +1,12 @@
 <?php
 
 use Kirby\Cms\App;
-use Kirby\Kql\Kql;
 
 @include_once __DIR__ . '/vendor/autoload.php';
 
-class_alias('Kirby\Kql\Kql', 'Kql');
-class_alias('Kirby\Kql\Interceptor', 'Kirby\Kql\Interceptors\Interceptor');
-
-function kql($input, $model = null)
-{
-	return Kql::run($input, $model);
-}
+require_once 'extensions/aliases.php';
+require_once 'extensions/helpers.php';
 
 App::plugin('getkirby/kql', [
-	'api' => [
-		'routes' => function ($kirby) {
-			return [
-				[
-					'pattern' => 'query',
-					'method'  => 'POST|GET',
-					'auth'    => $kirby->option('kql.auth') === false ? false : true,
-					'action'  => function () {
-						$result = Kql::run(get());
-
-						return [
-							'code'   => 200,
-							'result' => $result,
-							'status' => 'ok',
-						];
-					}
-				]
-			];
-		}
-	]
+	'api' => require_once 'extensions/api.php'
 ]);
